move initial connection room state from music room context to room component
"yes. It's not as expensive and janky as the first render where your app builds for the first time because compares new render with old and only updates differences. And if there are none, you don't.. See anything. But the expensive part here is calculating the difference. You don't want to permanently compare your whole app." - const reactive = "Gerrit" from r/webdev discord